backend/requirements.txt
fastapi
uvicorn==0.16.0
pydantic==2.10.2
sqlalchemy
asyncpg
redis
celery[redis]>=5.3,<6
boto3
python-multipart
pydantic-settings==2.6.1
passlib
httpx
python-jose[cryptography]